eport_2nd_task/requirements.txt
# Production dependencies
fastapi>=0.110.0
alembic>=1.13.1
uvicorn>=0.28.0
pydantic>=2.6.4
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
requests>=2.31.0
python-dotenv>=1.0.1
loguru>=0.7.2
psycopg2>=2.9.9
hypercorn>=0.16.0
asgi-correlation-id>=4.3.1
asyncpg>=0.29.0
greenlet>=3.0.3
httpx>=0.27.0
python-jose>=3.3.0
sqlalchemy>=2.0.36
sqlacodegen==3.0.0rc5
pydantic-settings>=2.2.1
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

